feat: pglite as database backend

+++ src/node/os/is-process-alive.ts
@@ -1,4 +1,8 @@
 export function isProcessAlive(pid: number) {
+  if (!Number.isInteger(pid) || pid <= 0) {
+    return false;
+  }
+
   try {
     process.kill(pid, 0);
     return true;
